+++ packages/fcl/src/current-user/index.js
@@ -109,14 +109,6 @@ async function authenticate({service, redir = false} = {}) {
     const opts = {redir}
     const discoveryService = await getDiscoveryService()
     const refreshService = serviceOfType(user.services, "authn-refresh")
-    const suppressRedirWarning = await config.get("fcl.warning.suppress.redir")
-
-    if (redir && !suppressRedirWarning) {
-      console.warn(
-        `You are manually enabling a very experimental feature that is not yet standard, use at your own risk.
-         You can disable this warning by setting fcl.warning.suppress.redir to true in your config`
-      )
-    }
     
     invariant(
       service || discoveryService.endpoint,